move util to prefix
